requirements.txt
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
kubernetes==29.0.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
psutil==5.9.8
httpx==0.27.0
